Move get_s3_client() to utils namespace
