+++ inc/namespace.php
@@ -8,7 +8,6 @@
 namespace Altis\Analytics;
 
 use Altis\Analytics\Utils;
-use Aws\S3\S3Client;
 use DateInterval;
 use DateTime;
 use Exception;
@@ -401,61 +400,6 @@ function delete_old_indexes() {
 }
 
 /**
- * Get a configured instance of S3 Client class.
- *
- * @param array $args Additional arguments to use with the client constructor.
- *
- * @return S3Client|null
- */
-function get_s3_client( array $args = [] ) : ? S3Client {
-
-	// These constants are required to continue.
-	if ( ! defined( 'ALTIS_ANALYTICS_PINPOINT_BUCKET_ARN' ) ) {
-		return null;
-	}
-	if ( ! defined( 'ALTIS_ANALYTICS_PINPOINT_BUCKET_REGION' ) ) {
-		return null;
-	}
-
-	$params = array_merge( [
-		'version' => '2006-03-01',
-		'region' => ALTIS_ANALYTICS_PINPOINT_BUCKET_REGION,
-	], $args );
-
-	// Add defined credentials if available.
-	if ( defined( 'ALTIS_ANALYTICS_S3_KEY' ) && defined( 'ALTIS_ANALYTICS_S3_SECRET' ) ) {
-		$params['credentials'] = [
-			'key' => ALTIS_ANALYTICS_S3_KEY,
-			'secret' => ALTIS_ANALYTICS_S3_SECRET,
-		];
-	}
-
-	// Allow overriding the S3 endpoint.
-	if ( defined( 'ALTIS_ANALYTICS_S3_ENDPOINT' ) ) {
-		$params['endpoint'] = ALTIS_ANALYTICS_S3_ENDPOINT;
-	}
-
-	/**
-	 * Filter the Analytics S3 client params.
-	 *
-	 * @param array $params The parameters used to instantiate the S3Client object.
-	 */
-	$params = apply_filters( 'altis.analytics.s3_client_params', $params );
-
-	$client = new S3Client( $params );
-
-	/**
-	 * Filter the S3 client used by the AWS Analytics plugin.
-	 *
-	 * @param Aws\S3\S3Client $client The S3Client object.
-	 * @param array $params The default params passed to the client.
-	 */
-	$client = apply_filters( 'altis.analytics.s3_client', $client, $params );
-
-	return $client;
-}
-
-/**
  * Clean up S3 long term data storage.
  *
  * @return void
@@ -463,7 +407,7 @@ function get_s3_client( array $args = [] ) : ? S3Client {
 function clean_s3_store() : void {
 
 	// Get S3 client.
-	$client = get_s3_client();
+	$client = Utils\get_s3_client();
 	if ( ! $client ) {
 		return;
 	}
